requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.30.0
datasets>=2.14.0,<4.0.0
PyYAML>=6.0
Pillow>=9.0.0
accelerate>=0.20.0
# Metrics (EM, BLEU, CIDEr, ROUGE-L)
rouge-score>=0.1.2
sacrebleu>=2.0.0
pycocoevalcap>=1.2
numpy>=1.20.0
